git-svn-id: svn://192.168.1.10/public/CoffeeMud@5087 0d6f1817-ed0e-0410-87c9-987e46238f29

+++ com/planet_ink/coffee_mud/Abilities/Common/LockSmith.java
@@ -36,6 +36,7 @@ public class LockSmith extends CraftingSkill
 	private Environmental workingOn=null;
 	private boolean messedUp=false;
 	private boolean boltlock=false;
+    private boolean delock=false;
 
 	public void unInvoke()
 	{
@@ -49,6 +50,7 @@ public class LockSmith extends CraftingSkill
 					if(messedUp)
 						commonTell(mob,"You've ruined "+building.name()+"!");
 					else
+                    if(!delock)
 						mob.location().addItemRefuse(building,Item.REFUSE_PLAYER_DROP);
 				}
 				building=null;
@@ -84,7 +86,7 @@ public class LockSmith extends CraftingSkill
 				{
 					if(workingOn instanceof Exit)
 					{
-						if(!((Exit)workingOn).hasALock())
+						if((delock)||(!((Exit)workingOn).hasALock()))
 						{
 							int dir=-1;
 							for(int d=0;d<Directions.DIRECTIONS_BASE.length;d++)
@@ -92,7 +94,10 @@ public class LockSmith extends CraftingSkill
 								{dir=Directions.DIRECTIONS_BASE[d]; break;}
 							if((messedUp)||(dir<0))
 							{
-								commonTell(mob,"You've ruined the lock.");
+                                if(delock)
+                                    commonTell(mob,"You've failed to remove the lock.");
+                                else
+    								commonTell(mob,"You've ruined the lock.");
 								building=null;
 								unInvoke();
 							}
@@ -102,7 +107,7 @@ public class LockSmith extends CraftingSkill
     							Room room2=mob.location().getRoomInDir(dir);
     							((Exit)workingOn).baseEnvStats().setLevel(mob.envStats().level());
     							((Exit)workingOn).recoverEnvStats();
-    							((Exit)workingOn).setDoorsNLocks(true,false,true,true,true,true);
+    							((Exit)workingOn).setDoorsNLocks(true,false,true,!delock,!delock,!delock);
     							if(building instanceof Key)
                                 {
                                     if(((Key)building).getKey().length()==0)
@@ -111,13 +116,13 @@ public class LockSmith extends CraftingSkill
                                 }
     							CMClass.DBEngine().DBUpdateExits(mob.location());
     							if((exit2!=null)
-    							   &&(!boltlock)
-    							   &&(exit2.hasADoor())
-    							   &&(exit2.isGeneric())
-    							   &&(room2!=null))
+							    &&(!boltlock)
+							    &&(exit2.hasADoor())
+							    &&(exit2.isGeneric())
+							    &&(room2!=null))
     							{
     								exit2.baseEnvStats().setLevel(mob.envStats().level());
-    								exit2.setDoorsNLocks(true,false,true,true,true,true);
+    								exit2.setDoorsNLocks(true,false,true,!delock,!delock,!delock);
     								if(building instanceof Key)
     									exit2.setKeyName(((Key)building).getKey());
     								CMClass.DBEngine().DBUpdateExits(room2);
@@ -128,17 +133,20 @@ public class LockSmith extends CraftingSkill
 					else
 					if(workingOn instanceof Container)
 					{
-						if(!((Container)workingOn).hasALock())
+						if(delock||(!((Container)workingOn).hasALock()))
 						{
 							if(messedUp)
 							{
-								commonTell(mob,"You've ruined the lock.");
+                                if(delock)
+                                    commonTell(mob,"You've failed to remove the lock.");
+                                else
+    								commonTell(mob,"You've ruined the lock.");
 								building=null;
 								unInvoke();
 							}
                             else
                             {
-    							((Container)workingOn).setLidsNLocks(true,false,true,true);
+    							((Container)workingOn).setLidsNLocks(true,false,!delock,!delock);
     							if(building instanceof Key)
                                 {
                                     if(((Key)building).getKey().length()==0)
@@ -158,7 +166,7 @@ public class LockSmith extends CraftingSkill
 	{
 		if(commands.size()==0)
 		{
-			commonTell(mob,"Locksmith what or where? Enter the name of a container or door direction. Put the word \"boltlock\" in front of the door direction to make a one-way lock.");
+			commonTell(mob,"Locksmith what or where? Enter the name of a container or door direction. Put the word \"boltlock\" in front of the door direction to make a one-way lock.  Put the word \"delock\" in front of the door direction to remove the locks.");
 			return false;
 		}
         keyCode=""+Math.random();
@@ -175,6 +183,12 @@ public class LockSmith extends CraftingSkill
 			lboltlock=true;
 			commands.removeElementAt(0);
 		}
+        boolean ldelock=false;
+        if((commands.size()>0)&&("DELOCK".startsWith(((String)commands.firstElement()).toUpperCase())))
+        {
+            ldelock=true;
+            commands.removeElementAt(0);
+        }
 		String recipeName=Util.combine(commands,0);
 		int dir=Directions.getGoodDirectionCode(recipeName);
 		if(dir<0)
@@ -199,10 +213,13 @@ public class LockSmith extends CraftingSkill
 				commonTell(mob,"That door isn't built right -- it can't be modified.");
 				return false;
 			}
-			if(((Exit)workingOn).hasALock())
-				keyFlag=true;
-			else
-				woodRequired=5;
+            if(!ldelock)
+            {
+    			if(((Exit)workingOn).hasALock())
+    				keyFlag=true;
+    			else
+    				woodRequired=5;
+            }
 
 			Room otherRoom=(dir>=0)?mob.location().getRoomInDir(dir):null;
 			if((!CoffeeUtensils.doesOwnThisProperty(mob,mob.location()))
@@ -226,10 +243,13 @@ public class LockSmith extends CraftingSkill
 				commonTell(mob,"That just isn't built right -- it can't be modified.");
 				return false;
 			}
-			if(((Container)workingOn).hasALock())
-				keyFlag=true;
-			else
-				woodRequired=3;
+            if(!ldelock)
+            {
+    			if(((Container)workingOn).hasALock())
+    				keyFlag=true;
+    			else
+    				woodRequired=3;
+            }
 		}
 		else
 		{
@@ -237,17 +257,31 @@ public class LockSmith extends CraftingSkill
 			return false;
 		}
 
-		int[] pm={EnvResource.MATERIAL_METAL,EnvResource.MATERIAL_MITHRIL};
-		int[][] data=fetchFoundResourceData(mob,
-											woodRequired,"metal",pm,
-											0,null,null,
-											false,
-											0);
-		if(data==null) return false;
-		woodRequired=data[0][FOUND_AMT];
-		if(!super.invoke(mob,commands,givenTarget,auto,asLevel))
-			return false;
-		destroyResources(mob.location(),woodRequired,data[0][FOUND_CODE],0,null,0);
+        String itemName=null;
+        if(ldelock)
+        {
+            itemName="a broken lock";
+            keyFlag=false;
+            if(!super.invoke(mob,commands,givenTarget,auto,asLevel))
+                return false;
+        }
+        else
+        {
+    		int[] pm={EnvResource.MATERIAL_METAL,EnvResource.MATERIAL_MITHRIL};
+    		int[][] data=fetchFoundResourceData(mob,
+    											woodRequired,"metal",pm,
+    											0,null,null,
+    											false,
+    											0);
+    		if(data==null) return false;
+    		woodRequired=data[0][FOUND_AMT];
+            if(!super.invoke(mob,commands,givenTarget,auto,asLevel))
+                return false;
+    		destroyResources(mob.location(),woodRequired,data[0][FOUND_CODE],0,null,0);
+            itemName=(EnvResource.RESOURCE_DESCS[(data[0][FOUND_CODE]&EnvResource.RESOURCE_MASK)]+" key").toLowerCase();
+            itemName=Util.startWithAorAn(itemName);
+            building.setMaterial(data[0][FOUND_CODE]);
+        }
 		building=getBuilding(workingOn);
 		if(building==null)
 		{
@@ -256,8 +290,6 @@ public class LockSmith extends CraftingSkill
 		}
 		completion=15-((mob.envStats().level()-workingOn.envStats().level()));
 		if(keyFlag) completion=completion/2;
-		String itemName=(EnvResource.RESOURCE_DESCS[(data[0][FOUND_CODE]&EnvResource.RESOURCE_MASK)]+" key").toLowerCase();
-		itemName=Util.startWithAorAn(itemName);
 		building.setName(itemName);
 		startStr="<S-NAME> start(s) working on "+(keyFlag?"a key for ":"")+workingOn.name()+".";
 		displayText="You are working on "+(keyFlag?"a key for ":"")+workingOn.name();
@@ -267,7 +299,6 @@ public class LockSmith extends CraftingSkill
 		building.setDescription(itemName+". ");
 		building.baseEnvStats().setWeight(woodRequired);
 		building.setBaseValue(1);
-		building.setMaterial(data[0][FOUND_CODE]);
 		if(keyFlag)
 			building.baseEnvStats().setLevel(1);
 		else
@@ -288,6 +319,7 @@ public class LockSmith extends CraftingSkill
 		{
 			mob.location().send(mob,msg);
 			boltlock=lboltlock;
+            delock=ldelock;
 			beneficialAffect(mob,mob,asLevel,completion);
 		}
 		return true;
